Use token_tools
