+++ requirements-dev.txt
@@ -2,6 +2,7 @@
 pygments
 pyeditorconfig
 beartype
+token_tools
 
 # Testing
 pytest
