+++ salve/server.py
@@ -3,6 +3,7 @@ from multiprocessing.queues import Queue as GenericClassQueue
 from time import sleep
 
 from pyeditorconfig import get_config
+from token_tools import Token
 
 from .misc import (
     COMMANDS,
@@ -13,7 +14,6 @@ from .misc import (
     ResponseQueueType,
 )
 from .server_functions import (
-    Token,
     find_autocompletions,
     get_definition,
     get_highlights,
