+++ salve/server_functions/highlight/__init__.py
@@ -1,2 +1,1 @@
 from .highlight import get_highlights  # noqa: F401
-from .tokens import Token, generic_tokens  # noqa: F401
