+++ setup.py
@@ -14,7 +14,9 @@ setup(
     url="https://github.com/salve-org/salve",
     long_description=long_description,
     long_description_content_type="text/markdown",
-    install_requires=["pygments", "pyeditorconfig", "beartype"],
+    install_requires=open("requirements.txt", "r+")
+    .read()
+    .splitlines(keepends=False),
     python_requires=">=3.11",
     license="MIT license",
     classifiers=[
